backend/requirements.txt
mysql-connector-python
fastapi[all]
cryptography